// components/TouchableNFWrapper/types.ts
export interface TouchableNFWrapperProps {
  children: any;
  onPress?: () => void;
  backgroundFeedback?: string;
  height?: string | number;
  borderRadius?: number;
  width?: string | number;
  backgroundColor?: string;
  elevation?: number;
  marginBottom?: string | number;
  marginTop?: string | number;
  marginLeft?: string | number;
  marginRight?: string | number;
  padding?: number;
}